third_party/WebKit/LayoutTests/fast/css/atapply/at-apply-shorthands.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
#e {
    --foo: {
        border: 10px solid green;
        font: 20px serif;
    };
    @apply --foo;
}
</style>
<div id=e>
</div>
<script>
test(function(){
    assert_equals(getComputedStyle(e).border, "10px solid rgb(0, 128, 0)");
    assert_equals(getComputedStyle(e).font, "normal normal normal normal 20px / normal serif");
}, "Basic usage of @apply with shorthand properties");
</script>
